Add role main.yaml to gen MOK and enroll it.
The role will be called uki_config, and will allow generating a
machine owner key locally based on a user-provided common name for the
certificate. Future versions should allow users to bring their own MOK,
but this is sufficient for the first release.

This file also includes some dracut configuration for UKI signing, but
this is considered a part of the same commit since it depends on files,
names and paths configured at the MOK creation phase.

Note that this commit doesn't provide a working tasks file due to a
type error somewhere at the "Enroll MOK" step. This is actively being
debugged.