Replace vars_prompt with ansible.builtin.pause.
It seems if I want to prompt for input I need to do it at the start of
each play. This is not reasonable since we only need to prompt for the
MOK password if it is being enrolled for the first time. This was
replaced by calls to ansible.builtin.pause in this commit. This is
not nearly as clean, and might cause some problems down the road. The
better solution is to involve ansible vault, although we will always
need to prompt before rebooting.