Always become to check MOK lists.
mokutil seems to silently fail when privileges are not sufficient to
consult a MOK list.

A bug was discovered just prior to this commit where declining a
reboot and running the play again would result in the MOK enrollment
process being skipped. It is really important that we fully enroll a MOK
before changing any configuration files, or the host will fail to boot
on the next kernel-install. This would be a nightmare to debug so a big
goal of this module is making all of these changes atomic.