Fix cert extensions, fix flow control errors, rebuild UKI.
This commit changes the extensions of the certificates made by openssl to use
the certificate format, so .pem, .der, and .priv respectively (the latter is
mostly convention). There are also various flow control error fixes that should
prevent successive executions of the playbook to trigger changes when the
configuration is incomplete. New package dependencies will be installed to
support booting straight into a UKI after shim in accordance with the f40
change proposal document. Note that this method is still experimental. Finally,
a handler was added to rebuild the UKI if key configuration blocks report a
change. The use of a handler avoid rebuilding multiple times, and makes it so
we don't need to store the change state of the role in a separate variable.