Change arg structure, add UKI rebuild handler.
The new argument structure is a lot flatter, but drops the pesign
singing tool entirely. There is no good ansible support for NSS ans/or
pesign, and it offers limited ability to import external private keys.
There is a python-nss library, so a full module could be developed in
theory - but it is a few years stale now. Since OpenSSL support is
native to Ansible, we use that with sbsign as a signing tool instead.
This means the certificate and key formats are the same for the dracut
and ukify generators, and they could be moved outside of the generator
argument. We also drop the need to specify a certificate name. I still
think most people will be importing external certificates, so they can
customize them to their heart's desire before running the role. As long
as the path of the generated certificates is provided to the role, they
will not be overwritten.

This commit uses a handler to trigger UKI rebuilding only when a change
occurs, but this is a bit broken at the moment. It works on first run,
but if the UKI rebuild fails, the next playbook run will not trigger a
change, and the handler is never run. Before we can merge this, work
needs to be done on a backup/restore system to avoid writing a bad boot
configuration that stays unnoticed until a new kernel update is
released.