Fix control flow, revert changes on failure.
This commit fulfils our atomicity requirement at the expense of
idempotency. Changes are successfully reverted when the UKI build fails,
but the UKI is rebuilt whether a change to the configuration occured or
not. This will eventually improve, but requires some ugly change
checking so wasn't done here.

The current role also reports changes when it arguably shouldn't, like
when fetching a file from the remote host or restoring a same-play
backup. This is low hanging fruit and should be fixed before the merge.