Fix file backups, fix restore idempotency.
Instead of the previous backup strategy that uses a temporary directory
on the host, this commit uses the "slurp" module to store file contents
in memory. This fixes an issue where backup files were not stored in a
reliably accessible location for error handlers.

Instead of reporting change when restoring a backed up configuration
file, this role now reports no change. The timestamp of the file
actually does change, but this is not considered a change for the sake
of this role. On request from clients we can also copy file attributes
during the restore process to keep an exactly identical filesystem state
before the run and after the restore.