Move requirements file to the meta dir in role.
These recent commits are confusing I know, this is related to the
"official" ansible-lint github action not being able to find collections
that are supposed to be installed by default. I'm having a difficult
time understanding where the requirements file will be read, but that is
likely because I use a slightly unusual structure for role distribution?
This commit uses reference from https://docs.ansible.com/ansible/latest/galaxy/user_guide.html#installing-roles-and-collections-from-the-same-requirements-yml-file