More troubleshooting on dependency install.
This commit moves the requirements.yml file back to the root dir (yes, I
know I tried this already), and moves the test playbook to the playbooks
dir, which is a standard collection root dir. This is meant to see if
the playbook.yaml file is being linted without resolving dependencies,
whereas a role might not? Note that the docs at
https://ansible.readthedocs.io/projects/lint/usage/#dependencies-and-requirements
conflict with the docs referenced in previous commits. The ansible-lint
docs don't suggest that the requirements.yml file will be detected in
roles/<role>/meta/requirements.yml.