Add mok role and tests.
This commit does not have a working test setup, and has not updated the
galaxy.yml file to advertise the role. Right now we are having the same
issues discussed in the testing platform issue, where we don't have a
way to programatically test this.

Next line of inquiry will be ehther we can use the already-enrolled
fedora CA certificate as the "enrolled" cert and a generated one as the
"not enrolled cert."

This commit assumes that we want to keep enrollment functionality, but
this is not yet decided at a project level. This could be turned into a
role that fails if the provided certificate is not enrolled. This would
simplify it too, since auto-generation of certificates would not be
required.